requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
trl>=0.7.0
peft>=0.6.0
datasets>=2.14.0
accelerate>=0.24.0

# Vision and image processing
pillow>=10.0.0
sentence-transformers>=2.2.0

# Utilities
pyyaml>=6.0
numpy>=1.24.0

# Optional but recommended
bitsandbytes>=0.41.0  # For 8-bit training
wandb>=0.15.0  # For experiment tracking

